Add fail-proof redudancy
As it stands currently, if for some reason the rdm log doesn't work properly, it can break the entire gamemode. This is because the execution chain when a lua error happens gets terminated before the util.StartBleeding() function can be called.

By moving the LogDMG() function to the bottom of GM:PlayerTakeDamage() function, the gamemode will still function properly if for some reason the rdm logging ever experiences a lua error. It's an extremely bad coding practice to have separate elements be able to break each other.