Patch Shotgun Firerate Exploit
This pull request patches an exploit with the shotgun which allows the user to decrease the firerate. If they fire then quickly change weapons, it will revert the firerate down to 0.25s regardless how much time is remaining. 

This patch only sets the 0.25s if there's less than 0.25s remaining on the 'GetNextPrimaryFire()' delay.